I removed the composite index that checks for uniqueness of the community_type and supertype pair.  This index broke installation if the Oracle blocksize is less than 8KB (which is true by default on at least some platforms).
The unique constraint was unnecessary since community_type was already declared as the
table's primary key.  Because of this, only one row with a given community_type name
is possible, which means the composite key (community_type, supertype) is guaranteed
to be unique.

Removing the index will also speed community creation slightly, save a bit of
disk space, etc etc.

Note that the composite index was useless for lookups on the supertype alone
anyway (since community_type was the leftmost member of the composite key), and that
any queries on the type hierarchy ought to use the denormalized table defined
by acs-objects-metadata.sql anyway.

So the index seems entirely useless to me.