The after-instantiate callback was setting the site's master template to "dotlrn-master-custom".  This is wrong, it should only be set at install time so users can override it with their own custom template if they want.
While this could be done in an after-install callback, in .LRN we have tried
to do such things in the install.xml file, giving users one easy-to-read
place to modify the initial configuration of .LRN.

Indeed, the install.xml file was setting the site default master to
dotlrn-master, only to have the callback overwrite it.

So I changed the install.xml file to set the default master to dotlrn-master-custom.

This is still not as cleanly done as I'd wish (for the default Sloan theme,
dotlrn-master works perfectly well) but we want to push out 2.2.1 alpha.